+++ CHANGES.md
@@ -1,6 +1,7 @@
 master (unreleased)
 -------------------
 
+* Fix default Host header value. See #150. (@ixti)
 * Remove BearerToken authorization header. (@ixti)
 * Fix handling of chunked responses without Content-Length header. (@ixti)
 * Remove `HTTP::Request#method` and deprecate `HTTP::Request#__method__`